1: Rename and update imports of 'CreateForm' in 'ManageRoomsForm' / 2: Now it has 3 props to display 3 buttons with corresponding CRUD functions : create, modify, delete (only creat works for now)
